feat(PPDSC-2091): finish codemod
