+++ lib/codemod/src/utils/add-prop.js
@@ -1,26 +1,8 @@
-const isObject = obj =>
-  typeof obj === 'object' && !Array.isArray(obj) && obj !== null;
+const {isObject, objectToExpression} = require('./helpers');
 
 const getAttributeValue = (path, attribute) =>
   path.node.openingElement.attributes.find(
     ({type, name: {name}}) => type === 'JSXAttribute' && name === attribute,
-  );
-
-// This function converts an object to a JSX expression.
-const objectToExpression = (jscodeshift, obj) =>
-  jscodeshift.objectExpression(
-    Object.entries(obj).reduce(
-      (prev, [k, v]) => [
-        ...prev,
-        jscodeshift.objectProperty(
-          jscodeshift.identifier(k),
-          isObject(v)
-            ? objectToExpression(jscodeshift, v)
-            : jscodeshift.literal(v),
-        ),
-      ],
-      [],
-    ),
   );
 
 // This function builds a nested object from an array of props and a final value.
