+++ lib/codemod/src/v7.0.0/__tests__/caption-inset.test.js
@@ -41,7 +41,6 @@ describe('caption-inset', () => {
     `;
 
     actual = captionInsetTransform({source}, {jscodeshift});
-    console.log(actual);
     expect(actual).toEqual(expected);
   });
   test('replace caption-inset with caption with overrides without space-inset', () => {
@@ -63,7 +62,9 @@ describe('caption-inset', () => {
         overrides={{
             typographyPreset: "editorialCaption010",
             anotherThing: "hello",
-            stylePreset: "mono"
+            stylePreset: "mono",
+            paddingBlock: "space020",
+            paddingInline: "space020"
         }} />
     `;
 
@@ -71,17 +72,20 @@ describe('caption-inset', () => {
     expect(actual).toEqual(expected);
   });
 
-  test('replace caption-inset with caption without overrides ', () => {
+  test('replace caption-inset with caption without overrides', () => {
     source = dedent`
           import { CaptionInset, Button } from 'newskit';
           <CaptionInset />
       `;
 
     expected = dedent`
-          import { Caption, Button } from 'newskit';
-          <Caption />
+    import { Caption, Button } from 'newskit';
+    <Caption
+      overrides={{
+        paddingBlock: "space020",
+        paddingInline: "space020"
+      }} />
       `;
-
     actual = captionInsetTransform({source}, {jscodeshift});
     expect(actual).toEqual(expected);
   });
