+++ lib/codemod/src/v7.0.0/caption-inset.js
@@ -1,53 +1,5 @@
 const {replaceComponent} = require('../utils/component-utils');
-const _ = require('lodash');
-
-const getChunk = (file, start, end) => file.substring(start, end);
-
-const strToObject = str => eval(`(${str})`);
-
-const objectToExpression = (jscodeshift, obj) =>
-  jscodeshift.objectExpression(
-    Object.entries(obj).reduce(
-      (prev, [k, v]) => [
-        ...prev,
-        jscodeshift.objectProperty(
-          jscodeshift.identifier(k),
-          isObject(v)
-            ? objectToExpression(jscodeshift, v)
-            : jscodeshift.literal(v),
-        ),
-      ],
-      [],
-    ),
-  );
-
-const isObject = obj =>
-  typeof obj === 'object' && !Array.isArray(obj) && obj !== null;
-
-const updateOverrides = (root, j, componentName, propFn) => {
-  const source = root.toSource();
-
-  return root.findJSXElements(componentName).forEach(path => {
-    const attributes = path.node.openingElement.attributes;
-    attributes.forEach((node, index) => {
-      if (node.type === 'JSXAttribute' && node.name.name === 'overrides') {
-        const objectChunk = getChunk(
-          source,
-          node.value.expression.start,
-          node.value.expression.end,
-        );
-
-        const overridesObject = strToObject(objectChunk);
-        const newOverrides = propFn(overridesObject);
-
-        attributes[index] = j.jsxAttribute(
-          j.jsxIdentifier('overrides'),
-          j.jsxExpressionContainer(objectToExpression(j, newOverrides)),
-        );
-      }
-    });
-  });
-};
+const {updateOverrides} = require('../utils/update-overrides');
 
 const map = {
   spaceInset010: {
@@ -93,14 +45,24 @@ module.exports = function transformer(file, api) {
   const root = jscodeshift(file.source);
 
   updateOverrides(root, jscodeshift, 'CaptionInset', overrides => {
+    const defaultPadding = {
+      paddingBlock: 'space020',
+      paddingInline: 'space020',
+    };
+
+    // when there is no overrides prop
+    // we want to add some defaults
+    if (!overrides) {
+      return defaultPadding;
+    }
+
     const {spaceInset, ...otherOverrides} = overrides;
 
     // does not have spaceInset so we add manually
     if (!spaceInset) {
       return {
         ...overrides,
-        paddingBlock: 'space020',
-        paddingInline: 'space020',
+        ...defaultPadding,
       };
     }
